Refactor out the need for RestUtils class
