+++ yamcs-core/src/main/java/org/yamcs/web/rest/archive/ArchiveAlarmRestHandler.java
@@ -12,13 +12,12 @@ import org.yamcs.utils.TimeEncoding;
 import org.yamcs.web.HttpException;
 import org.yamcs.web.InternalServerErrorException;
 import org.yamcs.web.NotFoundException;
-import org.yamcs.web.rest.RestRequest;
 import org.yamcs.web.rest.RestHandler;
+import org.yamcs.web.rest.RestRequest;
+import org.yamcs.web.rest.RestRequest.IntervalResult;
 import org.yamcs.web.rest.RestResponse;
 import org.yamcs.web.rest.RestStreamSubscriber;
 import org.yamcs.web.rest.RestStreams;
-import org.yamcs.web.rest.RestUtils;
-import org.yamcs.web.rest.RestUtils.IntervalResult;
 import org.yamcs.web.rest.SqlBuilder;
 import org.yamcs.web.rest.mdb.MDBHelper;
 import org.yamcs.web.rest.mdb.MDBHelper.MatchResult;
@@ -62,7 +61,7 @@ public class ArchiveAlarmRestHandler extends RestHandler {
         int limit = req.getQueryParameterAsInt("limit", 100);
         
         SqlBuilder sqlb = new SqlBuilder(AlarmRecorder.TABLE_NAME);
-        IntervalResult ir = RestUtils.scanForInterval(req);
+        IntervalResult ir = req.scanForInterval();
         if (ir.hasInterval()) {
             sqlb.where(ir.asSqlCondition("triggerTime"));    
         }
@@ -72,7 +71,7 @@ public class ArchiveAlarmRestHandler extends RestHandler {
         if (triggerTime != TimeEncoding.INVALID_INSTANT) {
             sqlb.where("triggerTime = " + triggerTime);
         }
-        sqlb.descend(RestUtils.asksDescending(req, true));
+        sqlb.descend(req.asksDescending(true));
         
         ListAlarmsResponse.Builder responseb = ListAlarmsResponse.newBuilder();
         RestStreams.streamAndWait(req, sqlb.toString(), new RestStreamSubscriber(pos, limit) {
